chore: split styles into commons and specific style ts
