+++ src/app/views/templates/commonStyles.ts
@@ -1,0 +1,26 @@
+export const textStyles = {
+  'body-1': {
+    fontStyle: 'normal',
+    fontWeight: 400,
+    fontSize: '16px',
+    lineHeight: '24px',
+    letterSpacing: '-0.176px',
+    fontFeatureSettings: "'tnum' on, 'lnum' on, 'cv05' on",
+  },
+  h4: {
+    fontWeight: 500,
+    fontSize: '18px',
+    lineHeight: '24px',
+    letterSpacing: '-0.252px',
+    fontFeatureSettings: "'tnum' on, 'lnum' on, 'cv05' on",
+  },
+  h5: {
+    fontWeight: 600,
+    fontSize: '20px',
+    lineHeight: '28px',
+    letterSpacing: '-0.28px',
+    fontFeatureSettings: "'tnum' on, 'lnum' on, 'cv05' on",
+  },
+}
+
+export const colorBaseContentStrong = '#2C2E34'
