+++ src/app/views/templates/styles.ts
@@ -56,79 +56,3 @@ export const linkStyle: NonNullable<LinkProps['style']> = {
   wordBreak: 'break-all',
   color: '#445fcd',
 }
-
-export const textStyles = {
-  'body-1': {
-    fontStyle: 'normal',
-    fontWeight: 400,
-    fontSize: '16px',
-    lineHeight: '24px',
-    letterSpacing: '-0.176px',
-    fontFeatureSettings: "'tnum' on, 'lnum' on, 'cv05' on",
-  },
-  h4: {
-    fontWeight: 500,
-    fontSize: '18px',
-    lineHeight: '24px',
-    letterSpacing: '-0.252px',
-    fontFeatureSettings: "'tnum' on, 'lnum' on, 'cv05' on",
-  },
-  h5: {
-    fontWeight: 600,
-    fontSize: '20px',
-    lineHeight: '28px',
-    letterSpacing: '-0.28px',
-    fontFeatureSettings: "'tnum' on, 'lnum' on, 'cv05' on",
-  },
-}
-
-const colorBaseContentStrong = '#2C2E34'
-
-export const mainStyle = {
-  fontFamily: 'sans-serif',
-  backgroundColor: '#f8f9fd',
-}
-
-export const containerStyle = {
-  maxWidth: '100%',
-  width: '720px',
-  margin: 'auto',
-  padding: '20px',
-}
-
-export const sectionStyle = {
-  padding: '32px',
-  backgroundColor: '#ffffff',
-}
-
-export const headingTextStyle: NonNullable<LinkProps['style']> = {
-  ...textStyles['h4'],
-  color: colorBaseContentStrong,
-  fontWeight: 600,
-  lineHeight: '32px',
-  fontSize: '24px',
-}
-
-export const outcomeTextStyle: NonNullable<LinkProps['style']> = {
-  ...textStyles['h5'],
-  color: colorBaseContentStrong,
-}
-
-export const primaryTextStyle: NonNullable<LinkProps['style']> = {
-  ...textStyles['body-1'],
-  color: colorBaseContentStrong,
-  fontWeight: 700,
-}
-
-export const secondaryTextStyle: NonNullable<LinkProps['style']> = {
-  ...textStyles['body-1'],
-  color: '#474747',
-}
-
-export const questionMargin: NonNullable<LinkProps['style']> = {
-  marginBottom: '4px',
-}
-
-export const answerMargin: NonNullable<LinkProps['style']> = {
-  marginTop: '4px',
-}
